bot/BOTrequirements.txt
aiogram
sqlalchemy
python-dotenv
httpx
asyncpg
asyncio